[HER-173] Moved components under list directory
