+++ src/cameras/map/sidebar/content/list/hooks/use-list-style.js
@@ -1,5 +1,9 @@
 import { makeStyles } from '@material-ui/core'
 
-const root = { width: '80vw', overflow: 'auto', maxHeight: 300 }
+const root = {
+  width: '100%',
+  overflow: 'auto',
+  maxHeight: 300,
+}
 
 export const useListStyle = makeStyles({ root })
