+++ src/cameras/map/sidebar/content/list/hooks/use-message-props.js
@@ -1,4 +1,4 @@
-import { useStyles } from 'cameras/list/style'
+import { useStyles } from 'cameras/map/sidebar/content/list/styles'
 
 export const useMessageProps = () => ({
   container: true,
