+++ src/cameras/map/sidebar/content/list/item-ids.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { usePickContent } from 'cameras/list/hooks/use-pick-content'
+import { usePickContent } from 'cameras/map/sidebar/content/list/hooks/use-pick-content'
 
 export const ItemIds = () => {
   const Content = usePickContent()
