+++ src/cameras/map/sidebar/content/list/items/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import { Item } from 'cameras/list/items/item'
-import { useItems } from 'cameras/list/items/hooks/use-item'
+import { Item } from 'cameras/map/sidebar/content/list/items/item'
+import { useItems } from 'cameras/map/sidebar/content/list/items/hooks/use-item'
 
 export const Items = () => {
   const items = useItems()
