+++ src/cameras/map/sidebar/content/list/styles.js
